requirements.txt
Flask
gunicorn
pysensorpush
google-cloud-secret-manager
git+https://github.com/DrGFreeman/IFTTT-Webhook.git